Attach .mdf/.ldf if they exist in container
If the container and its network are deleted, then the databases will be
dropped. So when `docker compose up` runs again to create the
containers, then the `CREATE DATABASE` command won't work since there is
already a .mdf/.ldf file existent. This is why the `FOR ATTACH` command
is needed in this case if the files already exist.