Improve Multi-Instance Docker Support and DB Restoration Fix (#47)
This PR will accomplish two tasks:

1. The ability for multiple Optimizely Docker projects to be created
without storage conflicts in /var/opt/mssql/host_data
- This commit 7a2b9b56d48050f51e4f6f48b76f157aa233bfb1 will add a .env
file to each Docker project that specifies a DB_DIRECTORY. This
environment variable will specify a new directory on
/var/opt/mssql/host_data/ within the container so that the DB may
persist when adding other docker projects on the base image of
mssql/server:2019-latest.
    
3. The ability to restore pre-existing databases from .mdf & .ldf files
on `docker compose up`
- If the container and its network are deleted, then the databases will
be dropped. So when `docker compose up` runs again to create the
containers, then the plain `CREATE DATABASE` command won't work since
there is already a .mdf & .ldf file copied over. This commit
3ca4fc27fff44e0089b24a5fbd7ebd740ed462e1 will check if the .mdf file
exists. If so, it will add a `FOR ATTACH` command to the end of `CREATE
DATABASE` to restore the DB from .mdf and .ldf.